Convert figures to SVG
Could not make it with qmd files,
only with ipynb.
Some changes are redundant, because there was big balagan trying to make this work.
The problem might have been the cells also having the plt output (like axes) that is not just the figure.
So appending `;` to end of cell might have worked.